add server routing logics and a sample home page for Marketing portal; add README
