+++ README.md
@@ -1,0 +1,17 @@
+# sales-and-marketing
+
+A system architecture solution for a Sales and Marketing system for Recreation USA
+
+## Usage
+
+Download and install [Node.js](https://nodejs.org/en/)
+
+Clone project into a folder
+
+```shell
+npm install
+npm start
+```
+
+Server should be started on port 3000 and marketing home page should be available on
+http://localhost:3000/marketing
