+++ config.js
@@ -1,0 +1,5 @@
+'use strict';
+module.exports = {
+  hostname: '127.0.0.1',
+  port: 3000
+}
