+++ package.json
@@ -23,7 +23,12 @@
   },
   "homepage": "https://github.com/cs490-group5/sales-and-marketing#readme",
   "devDependencies": {
-    "koa": "^1.2.5",
+    "ejs": "^2.5.6",
+    "koa": "^2.0.0",
+    "koa-mount": "^1.3.0",
+    "koa-router": "^5.4.0",
+    "koa-static": "^2.1.0",
+    "koa-views": "^5.2.0",
     "lodash": "^4.17.4"
   }
 }
