--- views/marketing/app.js
@@ -1,2 +1,0 @@
-'use strict';
-// placeholder file
