+++ views/marketing/index.html
@@ -1,0 +1,53 @@
+<html>
+  <head>
+    <title>Marketing Portal</title>
+    <link rel="import" href="/assets/head-bootstrap.html">
+    <!-- <script src="/assets/head-bootstrap.html"></script> -->
+  </head>
+  <body>
+    <div class="jumbotron">
+      <div class="container">
+        <h1>Marketing Portal</h1>
+        <p>This is the internal portal containing all the marketing tools!</p>
+      </div>
+    </div>
+
+    <div class="container">
+      <div class="row">
+        <div class="col-md-4">
+          <h2>Web Analytics</h2>
+          <p>See the key marketing metrics and analysis</p>
+          <p><a class="btn btn-default" href="#" role="button">Work in Progress</a></p>
+        </div>
+        <div class="col-md-4">
+          <h2>Marketing Mix Modeling</h2>
+          <p>Discover the impact of different marketing activities and predict sales</p>
+          <p><a class="btn btn-default" href="#" role="button">Work in Progress</a></p>
+       </div>
+        <div class="col-md-4">
+          <h2>Marketing Automation</h2>
+          <p>Set up automated advertising campaigns</p>
+          <p><a class="btn btn-default" href="#" role="button">Work in Progress</a></p>
+        </div>
+      </div>
+
+      <div class="row">
+        <div class="panel panel-default">
+          <div class="panel-heading">
+            <h3 class="panel-title">(@TODO: to be removed) page content testing area</h3>
+          </div>
+          <div class="panel-body">
+            Params <br/>
+            <%- JSON.stringify(params) %> <br/>
+            DB item <br/>
+            <%- JSON.stringify(item) %> <br/>
+          </div>
+        </div>
+      </div>
+
+      <footer>
+        <p>&copy; 2017 JF</p>
+      </footer>
+    </div>
+  </body>
+</html>
